Separate folders for  container and presentational components
