+++ lib/react/cell.js
@@ -1,7 +1,7 @@
 /* @flow */
 
 import React from "react";
-import CellEditor from "./cell-editor-container";
+import CellEditor from "./containers/cell-editor";
 
 type Props = {
   cell: RowCell,
