+++ lib/react/containers/cell-editor.js
@@ -1,10 +1,10 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
-import CellEditor from "./cell-editor";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps } from "react-redux";
+import * as A from "../../action-creators";
+import CellEditor from "../presentationals/cell-editor";
 
 const mapDispatchToProps: MapDispatchToProps<Action, *, *> = (
   dispatch: Dispatch<Action>
