+++ lib/react/containers/cell.js
@@ -1,10 +1,10 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
-import Cell from "./cell";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps } from "react-redux";
+import * as A from "../../action-creators";
+import Cell from "../cell";
 
 const mapDispatchToProps: MapDispatchToProps<Action, *, *> = (
   dispatch: Dispatch<Action>
