+++ lib/react/containers/panel.js
@@ -1,12 +1,12 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
-import Panel from "./panel";
-import paginationSelector from "../reselectors/pagination";
-import newVisibleRowsSelector from "../reselectors/visible-rows";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps, MapStateToProps } from "react-redux";
+import * as A from "../../action-creators";
+import paginationSelector from "../../reselectors/pagination";
+import newVisibleRowsSelector from "../../reselectors/visible-rows";
+import Panel from "../panel";
 
 export default function newPanelContainer(columns: Columns) {
   const visibleRowsSelector = newVisibleRowsSelector(
