+++ lib/react/containers/resize-handle.js
@@ -1,11 +1,11 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
-import ResizeHandle from "./resize-handle";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps, MapStateToProps } from "react-redux";
+import * as A from "../../action-creators";
+import ResizeHandle from "../presentationals/resize-handle";
 
 const mapStateToProps: MapStateToProps<State, *, *> = (state: State) => {
   return {
