+++ lib/react/containers/row.js
@@ -1,10 +1,10 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
-import Row from "./row";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps } from "react-redux";
+import * as A from "../../action-creators";
+import Row from "../row";
 
 const mapDispatchToProps: MapDispatchToProps<Action, *, *> = (
   dispatch: Dispatch<Action>
