+++ lib/react/containers/scrollable-list.js
@@ -1,10 +1,10 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
-import ScrollableList from "./scrollable-list";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps, MapStateToProps } from "react-redux";
+import * as A from "../../action-creators";
+import ScrollableList from "../presentationals/scrollable-list";
 
 const mapStateToProps: MapStateToProps<State, *, *> = (state: State) => {
   return {
