+++ lib/react/containers/search.js
@@ -1,10 +1,10 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
-import Search from "./search";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps, MapStateToProps } from "react-redux";
+import * as A from "../../action-creators";
+import Search from "../presentationals/search";
 
 const mapStateToProps: MapStateToProps<State, *, *> = (state: State) => {
   return {
