+++ lib/react/containers/table-column.js
@@ -1,10 +1,10 @@
 /* @flow */
 
-import * as A from "../action-creators";
 import { connect } from "react-redux";
-import { TableColumn } from "./table-column";
 import type { Dispatch } from "redux";
 import type { MapDispatchToProps, MapStateToProps } from "react-redux";
+import * as A from "../../action-creators";
+import { TableColumn } from "../presentationals/table-column";
 
 const mapStateToProps: MapStateToProps<State, *, *> = (state: State) => {
   const sort = state.sifterResult.options.sort[0];
