+++ lib/react/panel.js
@@ -4,12 +4,12 @@ import classNames from "classnames";
 import React from "react";
 import ReactDOM from "react-dom";
 import { Observable } from "rxjs";
-import Loading from "./loading";
-import Row from "./row-container";
-import ResizeHandle from "./resize-handle-container";
-import ScrollableList from "./scrollable-list-container";
-import Search from "./search-container";
-import TableColumn from "./table-column-container";
+import Loading from "./presentationals/loading";
+import Row from "./containers/row";
+import ResizeHandle from "./containers/resize-handle";
+import ScrollableList from "./containers/scrollable-list";
+import Search from "./containers/search";
+import TableColumn from "./containers/table-column";
 
 type Props = {
   columnHeaders: ColumnHeader[],
