+++ lib/react/presentationals/cell-editor.js
