+++ lib/react/presentationals/resize-handle.js
