+++ lib/react/presentationals/search.js
@@ -1,9 +1,9 @@
 /* @flow */
 
-import Disposables from "../disposables";
 import React from "react";
 import ReactDOM from "react-dom";
 import { Subject } from "rxjs";
+import Disposables from "../../disposables";
 
 const IS_INPUT_ELEMENT_REGEX = /INPUT|ATOM-TEXT-EDITOR/;
