+++ lib/react/presentationals/table-column.js
