+++ lib/react/row.js
@@ -2,7 +2,7 @@
 
 import classNames from "classnames";
 import React from "react";
-import Cell from "./cell-container";
+import Cell from "./containers/cell";
 
 type Props = {
   row: NoteRow,
