+++ lib/session.js
@@ -11,7 +11,7 @@ import * as A from "./action-creators.js";
 import NotesFileFilter from "./notes-file-filter";
 import newRootReducer from "./reducers/index";
 import NotesCache from "./notes-cache";
-import newPanelContainer from "./react/panel-container";
+import newPanelContainer from "./react/containers/panel";
 import RestartSessionForNewConfigToTakeEffect from "./restart-session-for-new-config-to-take-effect";
 import ToggleAtomWindow from "./toggle-atom-window";
 import TogglePanel from "./toggle-panel";
